Move `AuxProofVerifier` to air crate
