+++ air/src/lib.rs
@@ -42,10 +42,11 @@ pub use options::{FieldExtension, ProofOptions};
 
 mod air;
 pub use air::{
-    Air, AirContext, Assertion, AuxRandElements, BoundaryConstraint, BoundaryConstraintGroup,
-    BoundaryConstraints, ConstraintCompositionCoefficients, ConstraintDivisor,
-    DeepCompositionCoefficients, EvaluationFrame, LagrangeConstraintsCompositionCoefficients,
-    LagrangeKernelBoundaryConstraint, LagrangeKernelConstraints, LagrangeKernelEvaluationFrame,
-    LagrangeKernelRandElements, LagrangeKernelTransitionConstraints, TraceInfo,
-    TransitionConstraintDegree, TransitionConstraints,
+    Air, AirContext, Assertion, AuxProofVerifier, AuxRandElements, BoundaryConstraint,
+    BoundaryConstraintGroup, BoundaryConstraints, ConstraintCompositionCoefficients,
+    ConstraintDivisor, DeepCompositionCoefficients, DefaultAuxProofVerifier, EvaluationFrame,
+    LagrangeConstraintsCompositionCoefficients, LagrangeKernelBoundaryConstraint,
+    LagrangeKernelConstraints, LagrangeKernelEvaluationFrame, LagrangeKernelRandElements,
+    LagrangeKernelTransitionConstraints, TraceInfo, TransitionConstraintDegree,
+    TransitionConstraints,
 };
